feat: added microvmcluster controller
Initial implementation of the controller for MicrovmCluster. There isn't
any infrastructure for the controller to provision so the role of the
controller is mostly to satisfy the contract with CAPI.

As we will be using kube-viop as the default solution for load-balancing
there is an assumption that the controlplane endpoint will be filled in
on the MicrovmCluster. We use this address to test if the control plane
nodes are up and if they are we set a condition to say the load balancer
is available.

Note: CAPI will report errors as we have a control plane endpoint
defined which it will start to use. The errors will stop once the
control plane nodes have been provisioned. We may need to conider an
intermediary field/struct for the load balancer and only populate the
controlplane endpoint when the nodes are available.

Signed-off-by: Richard Case <richard@weave.works>