fix: failure domains reporting from microvmcluster
The API definitions for the MicrovmMachine was incorrectly specify
**FailureDomain** which is a deprecated way to specify a failure domain.
This has been removed and the code now looks for failure domain on the
**Machine** instead as this is the correct way.

Declarative placement has been introduced at the **MicrovmCluster**
level with the ability to set a static list of host endpoints to use
when creating microvms. This relies on the failure domain logic to
assign a **Machine** to one of the host endpoints.

Signed-off-by: Richard Case <richard@weave.works>