fix: do not default mount point for volumes
Taken of the kubebuilder default for the mount point of volumes,
Previously it defaulted to `/` and so if there was more than 1 volume
specified without mount points set there would be collision of mount
points. It has been changed to a required field without a default to
force users to specify it.

Signed-off-by: Richard Case <richard@weave.works>