feat: set provider id earlier (#114)
* feat: Return mvm object from service.Create

This returns the microvm in the Create response so that we can use
details for subsequent actions, eg status checks and getting the UUID.

* feat: do not requeue after a create

Instead we can just switch on the state immediately (at that point it
will be the default `pending`) and update the cluster conditions before
requeueing.

* feat: set the provider ID earlier

At the old location the ID was not persisted before the state switch,
meaning that we would sometimes (in almost all cases) exit and requeue
before it was set. This lead to dupe mvms being created on multiple
hosts.