fix: save failureDomain on mvm spec
We previously assumed that the failureDomain would be set on the
Machine Spec and that we could use that to determine placement.
This is eventually the case, but is set after a create is complete.
If we don't want to have duplicate mvms created on every host, then we
need to save this on our own MicrovmMachine Spec.