feat: provider id changes & add microvm proxy
Changed the machine reconciliation so that it uses the standard format
for the provider id with uses a prefix. The prefix for this provider is
`microvm://`. The provider id is parsed when looking up the status of
the microvm.

Additionally, the ability to explicitly set a proxy server for the
microvm service has been added to the APIs. This is for scenarios where
you don't want to use the HTTP[S]_PROXY env vars.

Signed-off-by: Richard Case <richard@weave.works>