Go back to having just 1 secret
**User Value**

This commit adds a parameter on the MicrovmClusterSpec to configure mTLS
for flintlock host client connections.

For now, mTLS is always configured with no option for simple TLS.

Credentials must be stored in a tls secret which is created in the same
namespace as the MicroVMCluster. The name of the secret is then passed
to the `tlsSecretRef` field on the cluster spec. The presence of this
field will determine whether CAPMVM loads the credentials into client
calls. Otherwise all calls will be made insecurely.

```
---
apiVersion: infrastructure.cluster.x-k8s.io/v1alpha1
kind: MicrovmCluster
metadata:
  name: mvm-test
  namespace: default
spec:
  tlsSecretRef: mytlssecret
...
```

The secret must contain the following fields in its `data`: `"tls.crt"`
(the client certificate), `"tls.key"` (the client key), and `"ca.crt"` (the
certificate authority cert).

```
apiVersion: v1
kind: Secret
metadata:
  name: mytlssecret
  namespace: default # same as the cluster
type: kubernetes.io/tls
data:
  tls.crt: |
    MIIC2DCCAcCgAwIBAgIBATANBgkqh ...
  tls.key: |
    MIIC2DCCAcCgAwIBAgIBATANBgkqh ...
  ca.crt: |
    MIIC2DCCAcCgAwIBAgIBATANBgkqh ...
```

**Implementation**

The secretref is on the mvm cluster object and will be used (secret fetched,
values decoded, etc) by every mvm machine on every reconcile which
requires it. This feels a bit wasteful to me, but apparently this is the
way things are done. Sharing in-memory data between reconciliation loops
is not really done (unless you want to write the value to the status of
one object so that it can be read by another, which in this case we
really don't want to do).

One idea is to have a global var. Another, better, idea is to have a
pointer value added to both reconcilers' structs: the mvm cluster writing
the value and the mvm machines reading it. I decided to leave this for
now and keep things simple.

**Testing**

I have added units for the extraction of the data from the secret, but
not for the addition of creds to the client. Tests that the client is
actually built with and is using the correct things would be more at the
integration level. Any tests around the host clients here use fake
clients so adding coverage there would show us nothing.

One thing I could do is create a fake flintlock (which would basically be
a naiive implementation of a flintlock server, returning hardcoded
answers) which verifies that the client is built the way we want it to be.
The benefits of this is that 1) we get fast turnaround at an integration
level without having to run an e2e suite, 2) we actually have some
integration level tests at all.
The downside is that the fake flintlock implementation may drift from
the real one, and we end up building CAPMVM capability which the real
flintlock does not support. Admittedly the chance of this happening is
quite narrow given that fake server would only return what we tell it
to, but just calling it out.

My next step is to create e2es which test all the interactions
between CAPMVM and flintlock.