feat: Document that TLS secret is Opaque type
We realised that using a `tls` secret type and adding on a `ca.crt` key
would be annoying for users. With the `tls.crt` and `tls.key` fields,
users have the options to supply the values without the `START` and `END`
headers/footers, which are added back by the Secrets controller. For the
`ca.crt`, we would require that the headers are present (adding the pem
to the `CaPool` with fail otherwise) and we don't want to have to
process the bytes ourselves.
Users can create the other fields with or without headers, but one rule for
part of the data, and another rule for the rest is an irritating thing to
document so this just makes it easier.
Certmanager creates all data values with the headers present, so this
will work for either users in either case.