fix: eplicitly close grpc client connection
When calling flintlock we where not explicitly closing the client
connection to the grpc server. As a result a goroutine is kept alive for
each client request and the number continue to rise until the server
can't accept any more requests.

This change will explicitly call `Close` for the underlying grpc client
connection which helps to ensure server side resource usage is kept to a
minimum.

Signed-off-by: Richard Case <richard.case@outlook.com>