feat: Add MountPoint option to AdditionalVolumes
And update templates to use a dedicated binary image for `kernel.image`
and an additional volume to provide the modules.

In line with changes to our image-builder pipeline, which removes the
embedded kernel modules from the OS image, this feature adds an option
to include modules as an additional volume.

Old images will remain available to not break those using
earlier versions.