feat: Validate rootvol and kernel spec in webhook
The previous commit made these fields optional, but we need to verify
that at least some kernel and root volume information is set.

This validation will check whether the short syntax for kernel and root
volume config are set, and if not, will validate that long syntax fields
are provided. Errors are returned if not.

Also add new `OS_VERSION` and `KERNEL_VERSION` fields to templates.